Chore: Migrado funcionalidade de GerarCSV do MainForms para um GerarCSVViewControl
