+++ MoveInatorForms/Structures/DependencyInjection.cs
@@ -2,6 +2,7 @@
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using MoveInatorForms.Forms;
+using MoveInatorForms.Forms.Views;
 using MoveInatorForms.Services;
 using MoveInatorForms.Services.Interfaces;
 
@@ -11,16 +12,29 @@ namespace MoveInatorForms.Structures
     {
         public static HostApplicationBuilder Inject(this HostApplicationBuilder builder)
         {
-            builder.Services.AddSingleton<ICSVService, CSVService>();
-
-            builder.Services.AddTransient<MainForm>();
-
-            return builder;
+            return builder.InjectServices()
+                          .InjectForms();
         }
 
         public static HostApplicationBuilder InjectContext(this HostApplicationBuilder builder, string connectionString)
         {
             builder.Services.AddDbContext<MoveInatorContext>(opt => opt.UseSqlServer(connectionString));
+
+            return builder;
+        }
+
+        private static HostApplicationBuilder InjectForms(this HostApplicationBuilder builder)
+        {
+            builder.Services.AddTransient<MainForm>();
+
+            builder.Services.AddTransient<GerarCSVSimplesViewControl>();
+
+            return builder;
+        }
+
+        private static HostApplicationBuilder InjectServices(this HostApplicationBuilder builder)
+        {
+            builder.Services.AddSingleton<ICSVService, CSVService>();
 
             return builder;
         }
